split up js into four modules
